backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
pydantic>=2.8.0,<3.0.0
pandas>=2.0.0
numpy>=2.0.0
scikit-learn>=1.4.0
joblib>=1.4.0
torch>=2.0.0,<3.0.0
requests>=2.31.0
reportlab>=4.0.0
python-multipart>=0.0.20
pytest>=8.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
